Move default implementation to core.
